+++ CHANGELOG.rst
@@ -1,7 +1,9 @@
 1.9.0 (2020-02-20)
 ------------------
 
-* Add new unit: "barrel per second" (``bbl/s``)
+* New ``classmethod`` ``Array.FromScalars`` that creates an ``Array`` from a ``List[Scalar]``.
+* Add new unit: "barrel per second" (``bbl/s``).
+
 
 1.8.0 (2020-01-10)
 ------------------